requirements.txt
langchain
langchain-community
langchain-openai
chromadb
pandas
sentence-transformers
python-dotenv
langchain-groq
streamlit
